pyproject.toml
[tool.poetry]
name = "django-rest-framework"
version = "0.1.0"
description = ""
authors = ["Ataraksia <1509lisavi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.6"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
pillow = "^11.1.0"
black = "^25.1.0"
djangorestframework-simplejwt = "^5.5.0"
django-filter = "^25.1"
ipython = "^9.2.0"
drf-yasg = "^1.21.10"
stripe = "^12.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
